refactor: update types of build route
